requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
fastapi==0.95.1
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.29.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
